requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-autoslug==1.9.8
Pillow==9.3.0
sqlparse==0.4.3